Use shard and realm properties in web3 API (#10554)
* use shard and realm for all entity type objects
* add static reference for common properties
* update rest entity stake query to use encoded entity id

---------

+++ hedera-mirror-web3/src/main/java/com/hedera/services/store/models/FcTokenAllowanceId.java
@@ -71,8 +71,8 @@ public class FcTokenAllowanceId implements Comparable<FcTokenAllowanceId> {
     public String toString() {
         return MoreObjects.toStringHelper(this)
                 .omitNullValues()
-                .add("tokenNum", tokenNum.longValue())
-                .add("spenderNum", spenderNum.longValue())
+                .add("tokenNum", tokenNum)
+                .add("spenderNum", spenderNum)
                 .toString();
     }
